% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate_movements.R
\name{animate_movements}
\alias{animate_movements}
\title{Animate the movement of the subject during the trial}
\usage{
animate_movements(
  data,
  exclude_start_minutes = 0,
  exclude_end_minutes = 0,
  speed = 100
)
}
\arguments{
\item{data}{An organised shuttle-box dataframe with corrected core body temperature}

\item{exclude_start_minutes}{Exclusion of time from the start of the trial onwards, in minutes. Default is 0}

\item{exclude_end_minutes}{Exclusion of time from the end of the trial backwards, in minutes. Default is 0}

\item{speed}{Speed setting of animation, default is 100}
}
\value{
Animation of the subject during the trial
}
\description{
This function animates the movement of the subject during the trial
}
