% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_distance.R
\name{calc_distance}
\alias{calc_distance}
\title{Calculate the cumulative distance}
\usage{
calc_distance(data, pixel_to_cm = TRUE)
}
\arguments{
\item{data}{An organised shuttle-box dataframe with corrected core body temperature}

\item{pixel_to_cm}{if TRUE, converts coordinates with pixels as unit to coordinates with cm as unit. Default is TRUE}
}
\value{
A shuttle-box dataframe with distance calculated for each observation
}
\description{
This function calculates the distance covered for each observation during the trial.
}
