% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_project_data.R
\name{compile_project_data}
\alias{compile_project_data}
\title{Compile all trials into a single datafile}
\usage{
compile_project_data(data_read)
}
\arguments{
\item{data_read}{A list object containing the imported shuttle-box trials}
}
\value{
Dataframe of all shuttle-box trials in a shuttle-box project
}
\description{
This function compiles all shuttle-box trials in a list object, as imported by read_shuttlesoft, in a single dataframe
}
