% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect.R
\name{inspect}
\alias{inspect}
\title{Inspect imported shuttle-box data}
\usage{
inspect(data)
}
\arguments{
\item{data}{an imported shuttle-box data file}
}
\description{
This function inspects shuttle-box data to check whether all variables are present and formatted correctly
}
