% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distance.R
\name{plot_distance}
\alias{plot_distance}
\title{Plot the cumulative distance during the trial}
\usage{
plot_distance(data, exclude_start_minutes = 0, exclude_end_minutes = 0)
}
\arguments{
\item{data}{An organised shuttle-box dataframe with corrected core body temperature}

\item{exclude_start_minutes}{Exclusion of time from the start of the trial onwards, in minutes. Default is 0}

\item{exclude_end_minutes}{Exclusion of time from the end of the trial backwards, in minutes. Default is 0}
}
\value{
Plot of the cumulative distance over time
}
\description{
This function plots the cumulative distance covered over time
}
