% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_histogram.R
\name{plot_histogram}
\alias{plot_histogram}
\title{Plot a histogram of a chosen shuttle-box metric}
\usage{
plot_histogram(
  data,
  column,
  binwidth = 0.1,
  exclude_start_minutes = 0,
  exclude_end_minutes = 0
)
}
\arguments{
\item{data}{An organised shuttle-box dataframe with corrected core body temperature}

\item{column}{Select the name of the column with the variable the histogram needs to be plotted from}

\item{binwidth}{Binwidth of the histogram}

\item{exclude_start_minutes}{Exclusion of time from the start of the trial onwards, in minutes. Default is 0}

\item{exclude_end_minutes}{Exclusion of time from the end of the trial backwards, in minutes. Default is 0}
}
\value{
A histogram of a chosen shuttle-box metric
}
\description{
This function plots a histogram of the frequency of values observed for a chosen shuttle-box metric
}
