% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_interval.R
\name{plot_interval}
\alias{plot_interval}
\title{Plot the interval means for a selected column over time}
\usage{
plot_interval(
  data,
  column,
  interval_minutes = 10,
  exclude_start_minutes = 0,
  exclude_end_minutes = 0
)
}
\arguments{
\item{data}{An organised shuttle-box dataframe with corrected core body temperature}

\item{column}{The name of the column to plot an interval for}

\item{interval_minutes}{The number of minutes per interval, default is 10}

\item{exclude_start_minutes}{Exclusion of time from the start of the trial onwards, in minutes. Default is 0}

\item{exclude_end_minutes}{Exclusion of time from the end of the trial backwards, in minutes. Default is 0}
}
\value{
Plot of the interval means for a selected column over time
}
\description{
This function plots the  the interval means for a selected column over time
}
