% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_speed_coreT.R
\name{plot_speed_coreT}
\alias{plot_speed_coreT}
\title{Plot relation between velocity and core body temperature}
\usage{
plot_speed_coreT(data, exclude_start_minutes = 0, exclude_end_minutes = 0)
}
\arguments{
\item{data}{An organised shuttle-box dataframe with corrected core body temperature}

\item{exclude_start_minutes}{Exclusion of time from the start of the trial onwards, in minutes. Default is 0}

\item{exclude_end_minutes}{Exclusion of time from the end of the trial backwards, in minutes. Default is 0}
}
\value{
Plot of velocity versus core body temperature
}
\description{
This function plots the velocity of the subject against the core body temperature
}
