% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_T_gradient.R
\name{plot_T_gradient}
\alias{plot_T_gradient}
\title{Plot the temperatures in each side of the shuttlebox over time}
\usage{
plot_T_gradient(data, exclude_start_minutes = 0, exclude_end_minutes = 0)
}
\arguments{
\item{data}{An organised shuttle-box dataframe}

\item{exclude_start_minutes}{Exclusion of time from the start of the trial onwards, in minutes. Default is 0}

\item{exclude_end_minutes}{Exclusion of time from the end of the trial backwards, in minutes. Default is 0}
}
\value{
a plot with the temperatures in the warm and cold chambers
}
\description{
This function plots the temperature in the warm and cold chamber during the trial
}
