% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_coreT.R
\name{calc_coreT}
\alias{calc_coreT}
\title{Calculate the core body temperature}
\usage{
calc_coreT(data)
}
\arguments{
\item{data}{an organised shuttle-box data file, as output by file_prepare}
}
\value{
a data frame with corrected core body temperature values
}
\description{
This function calculates the core body temperature changes during the trial
}
