% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_gravitation.R
\name{calc_gravitation}
\alias{calc_gravitation}
\title{Calculate the gravitation period}
\usage{
calc_gravitation(
  data,
  exclude_start_minutes = 0,
  exclude_end_minutes = 0,
  exclude_acclimation = F,
  print_results = T
)
}
\arguments{
\item{data}{An organised shuttle-box dataframe with corrected core body temperature}

\item{exclude_start_minutes}{Exclusion of time from the start of the trial onwards, in minutes. Default is 0}

\item{exclude_end_minutes}{Exclusion of time from the end of the trial backwards, in minutes. Default is 0}

\item{exclude_acclimation}{Exclude the acclimation period from variable calculation, default = F}

\item{print_results}{Print the results, default is TRUE}
}
\value{
the temperature preference
}
\description{
This function calculates the gravitation period using a segmented regression with a single breakpoint. The breakpoint represents the moment where temperature preference was reached. The time until that point is the gravitation time
}
