% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_project_results.R
\name{calc_project_results}
\alias{calc_project_results}
\title{Calculate shuttle-box metrics for all shuttle-box trials}
\usage{
calc_project_results(
  data_read,
  calculate_distance = F,
  pixel_to_cm = T,
  exclude_acclimation = F,
  exclude_start_minutes = 0,
  exclude_end_minutes = 0,
  Tpref_method = "median",
  Tavoid_percentiles = c(0.05, 0.95),
  textremes_threshold = expression(0.2 * (max(df$max_T) - max(df$min_T))),
  core_T_variance_type = "std_error"
)
}
\arguments{
\item{data_read}{A list object containing the imported shuttle-box trials}

\item{calculate_distance}{Boolean to determine whether to calculate the distance according to calculate_distance, default is FALSE}

\item{pixel_to_cm}{Specify whether to convert coordinate units in pixels to coordinate units in cm if calculate_distance is TRUE, default is TRUE}

\item{exclude_acclimation}{Exclude the acclimation period from variable calculation, default = F}

\item{exclude_start_minutes}{Exclusion of time from the start of the trial onwards, in minutes. Default is 0}

\item{exclude_end_minutes}{Exclusion of time from the end of the trial backwards, in minutes. Default is 0}

\item{Tpref_method}{The method used for calculation of temperature preference ("median", "mean", "mode"). Default is "median"}

\item{Tavoid_percentiles}{The lower and upper percentile for lower and upper avoidance temperature calculation resp. Default is c(0.05, 0.95)}

\item{textremes_threshold}{Definition of the extreme temperature range. Default is 20\% of the temperature range: expression(0.2*(max(data$max_T)-max(data$min_T)))}

\item{core_T_variance_type}{The method used for calculating the variance of core body temperature ("std_error", "std_deviation", "coeff_variation"). Default is "std_error".}
}
\value{
Dataframe containing shuttle-box metrics for all trials in the project
}
\description{
This function calculates shuttle-box metrics for all trials in a list object, as imported by read_shuttlesoft
}
