% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_shuttles.R
\name{calc_shuttles}
\alias{calc_shuttles}
\title{Calculate the number of shuttles}
\usage{
calc_shuttles(
  data,
  exclude_start_minutes = 0,
  exclude_end_minutes = 0,
  exclude_acclimation = F,
  print_results = T
)
}
\arguments{
\item{data}{An organised shuttle-box dataframe with corrected core body temperature}

\item{exclude_start_minutes}{Exclusion of time from the start of the trial onwards, in minutes. Default is 0}

\item{exclude_end_minutes}{Exclusion of time from the end of the trial backwards, in minutes. Default is 0}

\item{exclude_acclimation}{Exclude the acclimation period from variable calculation, default = F}

\item{print_results}{Print the results, default is TRUE}
}
\value{
The total number of shuttles
}
\description{
This function calculates the total number of times the subject shuttled during the trial
}
