% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_Tavoid.R
\name{calc_Tavoid}
\alias{calc_Tavoid}
\title{Calculate the avoidance temperatures}
\usage{
calc_Tavoid(
  data,
  percentiles = c(0.05, 0.95),
  exclude_start_minutes = 0,
  exclude_end_minutes = 0,
  exclude_acclimation = F,
  print_results = T
)
}
\arguments{
\item{data}{An organised shuttle-box dataframe with corrected core body temperature}

\item{percentiles}{The lower and upper percentile for lower and upper avoidance temperature calculation resp. Default is c(0.05, 0.95)}

\item{exclude_start_minutes}{Exclusion of time from the start of the trial onwards, in minutes. Default is 0}

\item{exclude_end_minutes}{Exclusion of time from the end of the trial backwards, in minutes. Default is 0}

\item{exclude_acclimation}{Exclude the acclimation period from variable calculation, default = F}

\item{print_results}{Print the results, default is TRUE}
}
\value{
the upper and lower avoidance temperature
}
\description{
This function calculates the upper and lower avoidance temperature for the trial
}
