% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_tot_distance.R
\name{calc_tot_distance}
\alias{calc_tot_distance}
\title{Calculate the total distance}
\usage{
calc_tot_distance(
  data,
  exclude_start_minutes = 0,
  exclude_end_minutes = 0,
  exclude_acclimation = FALSE,
  print_results = TRUE
)
}
\arguments{
\item{data}{An organised shuttle-box dataframe with corrected core body temperature}

\item{exclude_start_minutes}{Exclusion of time from the start of the trial onwards, in minutes. Default is 0}

\item{exclude_end_minutes}{Exclusion of time from the end of the trial backwards, in minutes. Default is 0}

\item{exclude_acclimation}{Exclude the acclimation period from variable calculation, default = F}

\item{print_results}{Print the results, default is TRUE}
}
\value{
the total distance covered during the trial
}
\description{
This function calculates the total distance covered during the trial
}
