% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_Tpref.R
\name{calc_Tpref}
\alias{calc_Tpref}
\title{Calculate the temperature preference}
\usage{
calc_Tpref(
  data,
  method = c("median", "mean", "mode"),
  exclude_acclimation = F,
  exclude_start_minutes = 0,
  exclude_end_minutes = 0,
  print_results = T
)
}
\arguments{
\item{data}{An organised shuttle-box dataframe with corrected core body temperature}

\item{method}{The method used for calculation of temperature preference ("median", "mean", "mode"). Default is "median".}

\item{exclude_acclimation}{Exclude the acclimation period from variable calculation, default = F}

\item{exclude_start_minutes}{Exclusion of time from the start of the trial onwards, in minutes. Default is 0}

\item{exclude_end_minutes}{Exclusion of time from the end of the trial backwards, in minutes. Default is 0}

\item{print_results}{Print the results, default is TRUE}
}
\value{
the temperature preference
}
\description{
This function calculates the temperature preference for the trial
}
