% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_track_accuracy.R
\name{calc_track_accuracy}
\alias{calc_track_accuracy}
\title{Calculate the tracking accuracy}
\usage{
calc_track_accuracy(
  data,
  exclude_start_minutes = 0,
  exclude_end_minutes = 0,
  exclude_acclimation = F,
  print_results = T
)
}
\arguments{
\item{data}{An organised shuttle-box dataframe with corrected core body temperature}

\item{exclude_start_minutes}{Exclusion of time from the start of the trial onwards, in minutes. Default is 0}

\item{exclude_end_minutes}{Exclusion of time from the end of the trial backwards, in minutes. Default is 0}

\item{exclude_acclimation}{Exclude the acclimation period from variable calculation, default = F}

\item{print_results}{Print the results, default is TRUE}
}
\value{
the tracking accuracy
}
\description{
This function calculates the proportion of time the subject was tracked accurately during the trial
}
