% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_matrix.R
\name{correlation_matrix}
\alias{correlation_matrix}
\title{Plot a correlation matrix of selected shuttle-box metrics}
\usage{
correlation_matrix(
  data,
  columns,
  exclude_start_minutes = 0,
  exclude_end_minutes = 0,
  interval_minutes = 10
)
}
\arguments{
\item{data}{An organised shuttle-box dataframe with corrected core body temperature}

\item{columns}{Specify the columns to be used in plotting the correlation matrix}

\item{exclude_start_minutes}{Exclusion of time from the start of the trial onwards, in minutes. Default is 0}

\item{exclude_end_minutes}{Exclusion of time from the end of the trial backwards, in minutes. Default is 0}

\item{interval_minutes}{Specify the interval length in minutes, default is 10}
}
\value{
Correlation matrix of selected shuttle-box metrics
}
\description{
This function plots the number of missing tracks per interval
}
