% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_prepare.R
\name{file_prepare}
\alias{file_prepare}
\title{Prepare shuttle-box data}
\usage{
file_prepare(data)
}
\arguments{
\item{data}{a raw shuttle-box dataframe, for example as output by read_shuttlesoft}
}
\value{
an organised shuttle-box dataframe, ready for calculation steps
}
\description{
This prepares shuttle-box data by adding the variables time_h and time_s (time in hours and time in seconds), correcting the timeseries in case the trial continues past midnight, and delimiting the acclimmation period using trial_start
}
