% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca}
\alias{pca}
\title{Perform a PCA on the project data and calculate outliers}
\usage{
pca(
  data,
  mahalanobis_th = 0.7,
  dbscan_th = 1,
  print_labels = TRUE,
  id_col = "fileID",
  var_col = "Tpref",
  biplot_variables = T
)
}
\arguments{
\item{data}{(Subset of) project data containg shuttle-box metrics, as output by calc_project_data}

\item{mahalanobis_th, }{threshold for mahalanobis distance outlier cutoff, default is 0.7}

\item{dbscan_th, }{threshold for DBSCAN clustering outlier cutoff, default is 1}

\item{print_labels}{Boolean to print labels for each datapoint, default is TRUE}

\item{id_col}{Column to be used for identifier labels in plots, default is "fileID"}

\item{var_col}{Column to plot against principal component, default is "Tpref"}

\item{biplot_variables}{Boolean to print overlaying of shuttle-box metrics on biplot}
}
\value{
List object containing PCA output
}
\description{
This function performs a Principal Component Analysis, and calculates outliers according to the Mahalanobis distance and DBSCAN clustering.
The function produces a list object containing the pca, pca loadings, pca scores, a table containing the outliers, a PCA biplot, a screeplot, a contribution plot and a plot of a chosen variable against the first principal component
}
