% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_coreT_histogram.R
\name{plot_coreT_histogram}
\alias{plot_coreT_histogram}
\title{Plot a histogram of core body temperatures}
\usage{
plot_coreT_histogram(
  data,
  bin_size = 0.1,
  exclude_start_minutes = 0,
  exclude_end_minutes = 0,
  exclude_acclimation = F
)
}
\arguments{
\item{data}{An organised shuttle-box dataframe with corrected core body temperature}

\item{bin_size}{Bin size of the histogram, default is 0.1}

\item{exclude_start_minutes}{Exclusion of time from the start of the trial onwards, in minutes. Default is 0}

\item{exclude_end_minutes}{Exclusion of time from the end of the trial backwards, in minutes. Default is 0}

\item{exclude_acclimation}{Exclude the acclimation period from variable calculation, default is TRUE}
}
\value{
Histogram of the core body temperatures during the trial
}
\description{
This function plots a histogram of different core body temperatures observed during the trial
}
