% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distance_vs_shuttles.R
\name{plot_distance_vs_shuttles}
\alias{plot_distance_vs_shuttles}
\title{Plot distance versus shuttles across project data}
\usage{
plot_distance_vs_shuttles(proj_data)
}
\arguments{
\item{proj_data}{Project data containg shuttle-box metrics, as output by calc_project_data}
}
\value{
Plot of distance versus shuttles across project data
}
\description{
This function plots the distance covered during the trial against the number of shuttles, across individual trials in the project data
}
