% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_limits_vs_distance.R
\name{plot_limits_vs_distance}
\alias{plot_limits_vs_distance}
\title{Plot distance versus time spent near limits across project data}
\usage{
plot_limits_vs_distance(proj_data)
}
\arguments{
\item{proj_data}{Project data containg shuttle-box metrics, as output by calc_project_data}
}
\value{
Plot of distance versus time near limits across project data
}
\description{
This function plots the distance covered during the trial against the time spent near the set temperature limits, across individual trials in the project data
}
