% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_limits_vs_shuttles.R
\name{plot_limits_vs_shuttles}
\alias{plot_limits_vs_shuttles}
\title{Plot time spent near the limits versus shuttles across project data}
\usage{
plot_limits_vs_shuttles(proj_data)
}
\arguments{
\item{proj_data}{Project data containg shuttle-box metrics, as output by calc_project_data}
}
\value{
Plot of limits versus shuttles across project data
}
\description{
This function plots the the time spent near the set temperature limits against the number of shuttles, across individual trials in the project data
}
