% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_T_segmented.R
\name{plot_T_segmented}
\alias{plot_T_segmented}
\title{Plot the core body temperature during the trial}
\usage{
plot_T_segmented(
  data,
  Tpref_method = "median",
  Tavoid_percentiles = c(0.05, 0.95),
  exclude_start_minutes = 0,
  exclude_end_minutes = 0,
  exclude_acclimation = TRUE,
  overlay_chamber_temp = T
)
}
\arguments{
\item{data}{An organised shuttle-box dataframe with corrected core body temperature}

\item{Tpref_method}{The method used for calculation of temperature preference ("median", "mean", "mode"). Default is "median".}

\item{Tavoid_percentiles}{The lower and upper percentile for lower and upper avoidance temperature calculation resp. Default is c(0.05, 0.95)}

\item{exclude_start_minutes}{Exclusion of time from the start of the trial onwards, in minutes. Default is 0}

\item{exclude_end_minutes}{Exclusion of time from the end of the trial backwards, in minutes. Default is 0}

\item{exclude_acclimation}{Exclude the acclimation period from variable calculation, default is TRUE}

\item{overlay_chamber_temp}{Boolean to determine whether or not to overlay chamber temperatures, default = TRUE}
}
\value{
plot with the core body temperature and key shuttle-box metrics
}
\description{
This function plots the core body temperature during the trial, along with temperature preference, avoidance, and gravitation time.
}
