% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_shuttlesoft.R
\name{read_shuttlesoft}
\alias{read_shuttlesoft}
\title{Import shuttlesoft shuttle-box data}
\usage{
read_shuttlesoft(file, metadata, multidat = F)
}
\arguments{
\item{file}{A string containing the directory of the file to be loaded.}

\item{metadata}{A metadata file containing the variables file_name, initial_temp, mass, a_value and b_value}

\item{multidat}{A parameter that specifies whether this considers a single trial or multiple trials, default is FALSE}
}
\value{
a raw shuttle-box dataframe
}
\description{
This function imports .txt files output by shuttlesoft
}
