% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShuttleboxR-package.R
\docType{package}
\name{ShuttleboxR-package}
\alias{ShuttleboxR}
\alias{ShuttleboxR-package}
\title{ShuttleboxR: An R Package For Data Exploration In Shuttle-Box Temperature Experiments}
\description{
ShuttleboxR is an R package designed to streamline the workflow between data collection and analysis. ShuttleboxR provides functions for data import, organization, calculation, and inspection, offering a standardized approach to data handling. It also helps identify flawed methodologies or trials, improving the reliability of shuttle-box experiments.
}
\author{
\strong{Maintainer}: Pieter Riesenkamp \email{p.b.riesenkamp@student.rug.nl}

}
\keyword{internal}
